pyproject.toml
[project]
name = "CATastrophe"
version = "0.1.0"
description = "AI-powered chatbot to aid people in case of emergencies and natural disasters."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.38.0",
    "ipykernel>=6.30.1",
    "langfuse>=3.8.1",
    "openinference-instrumentation-google-genai>=0.1.8",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.4",
]